refactor: same treatment for the RichEditor. Splitted in vairable/simple component.
