+++ src/components/Hero.ts
@@ -1,4 +1,4 @@
-import { TextInput, TagsInput, defineGrid, Tabs, TabField, Textarea } from '@/lib/components/form-builder/fields';
+import { TextInput, TagsInput, defineGrid, Tabs, TabField, Textarea, RichEditor } from '@/lib/components/form-builder/fields';
 import type { Component } from '@/lib/components/form-builder/types';
 import { IconBrandGithub, IconLink } from '@tabler/icons-svelte';
 
@@ -20,11 +20,10 @@ mainTab.schema = [
         .required()
         .type("email")
         .placeholder('Enter the contact email')
-        .allowVariables(false), // Example: disable variables for this email field
-    Textarea('description')
+        .allowVariables(false),
+    RichEditor('description')
         .label('Description (no variables)')
         .placeholder('Enter description without variable support')
-        .allowVariables(false), // Example: disable variables for this textarea
 ];
 
 const buttonsTab = defineGrid(2, 4, { sm: 1, md: 2, lg: 2 });
